Init Dev 0005
Initial message passing has been implemented. This version has code
implemented which will display the login screen. Then when a user clicks
the login button a message will be sent to SYS of type loginRequest. SYS
then checks if anything is entered in the username or password login
fields and if so SYS sends a message to CND to have CND display the
DoctorGUIPanel. This sequence will need to be updated to check the
database values eventually but serves as a good initial example for
future work to be done. Note that at the end of each run() function the
class must call IO.getInbox().emptyReadMessages(); in order for this to
work properly. Other general updates have been made in order to
integrate everything together. Specifically the SYS run function, CND
run function and various arch functions have been updated.