Dev Temp 0005-1
Fixes are needed in the architecture or integration so that null pointer
exceptions stop occurring. Currently these occur if you hit the login
button without entering anything, then exiting the window that pops up,
and repeat this process multiple times. Then if you click the forgot
username or forgot password links you will eventually get a null pointer
exception. This is occurring in the emptyreadMessages function as well
as in the getNewMessages function.