Code PCR 0004d
Implemented full capabilities to navigate menus for doctors, patients,
nurses, and MAs. Also added multiple message types to be updated in SYS
at a later time. These messages have to do with updating patients and
displaying patient info through CND. I also inmplemented a sysGoToMenu
command which is used when pressing the cancel buttons in the various
pages. Depending on the permissions of the user the cancel button will
return them to the proper menu panel.